pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "axoden"
authors = [{name = "Raquel Adaia Sandoval Ortega", email = "raqueladaia@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
"requires-python" = ">= 3.8"
dependencies = [
    "tqdm >= 4.0.0",
    "numpy >= 1.16.0",
    "pandas >= 2.0.0",
    "seaborn >= 0.10.0",
    "matplotlib >= 3.0.0",
    "pillow >= 10.0.0",
    "scikit-image >= 0.20.0",
]

[tool.setuptools.dynamic]
version = {attr = "axoden.__vesion__"}


[project.optional-dependencies]
gui = [
    "streamlit >= 1.30.0",
    "streamlit-pdf-viewer >= 0.0.10",
    "pypdf >= 4.0.0",
]

[tool.pytest.ini_options]
pythonpath = [
    '.'
]